Add efi and hyperv tests to windows OS with instance types
