+++ tests/infrastructure/instance_types/supported_os/test_windows_os.py
@@ -2,9 +2,8 @@ import logging
 
 import pytest
 
-from utilities.virt import (
-    running_vm,
-)
+from utilities.guest_support import assert_windows_efi, check_vm_xml_hyperv, check_windows_vm_hvinfo
+from utilities.virt import running_vm
 
 pytestmark = [pytest.mark.high_resource_vm, pytest.mark.tier3]
 
@@ -26,6 +25,18 @@ class TestCommonPreferenceWindows:
     @pytest.mark.polarion("CNV-12270")
     def test_start_vm(self, golden_image_windows_vm):
         running_vm(vm=golden_image_windows_vm)
+
+    @pytest.mark.dependency(depends=[f"{TESTS_CLASS_NAME}::start_vm"])
+    @pytest.mark.polarion("CNV-12387")
+    def test_efi_secureboot_enabled(self, golden_image_windows_vm):
+        assert_windows_efi(vm=golden_image_windows_vm)
+
+    @pytest.mark.sno
+    @pytest.mark.dependency(depends=[f"{TESTS_CLASS_NAME}::start_vm"])
+    @pytest.mark.polarion("CNV-12388")
+    def test_hyperv(self, golden_image_windows_vm):
+        check_vm_xml_hyperv(vm=golden_image_windows_vm)
+        check_windows_vm_hvinfo(vm=golden_image_windows_vm)
 
     @pytest.mark.dependency(depends=[f"{TESTS_CLASS_NAME}::create_vm"])
     @pytest.mark.polarion("CNV-12271")
