+++ tests/virt/cluster/common_templates/general/test_base_template.py
@@ -16,8 +16,7 @@ from ocp_resources.template import Template
 from pytest_testconfig import config as py_config
 
 from tests.os_params import FEDORA_LATEST_LABELS
-from tests.virt.cluster.common_templates.constants import HYPERV_FEATURES_LABELS_VM_YAML
-from utilities.constants import DATA_SOURCE_NAME, DATA_SOURCE_NAMESPACE, Images
+from utilities.constants import DATA_SOURCE_NAME, DATA_SOURCE_NAMESPACE, HYPERV_FEATURES_LABELS_VM_YAML, Images
 
 pytestmark = [pytest.mark.post_upgrade, pytest.mark.sno]
